First demo test written, getAllRestaurants method added and wishlist added to getCustomer/getCustomers methods

+++ src/main/java/com/example/AverageConnoisseurServerSideProject/restaurant/RestaurantService.java
@@ -1,5 +1,6 @@
 package com.example.AverageConnoisseurServerSideProject.restaurant;
 
+import com.example.AverageConnoisseurServerSideProject.exceptions.ResourceNotFound;
 import com.example.AverageConnoisseurServerSideProject.review.Review;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
@@ -12,25 +13,52 @@ public class RestaurantService {
 
     private RestaurantDAO restaurantDAO;
 
-    public RestaurantService(@Qualifier("postgres") RestaurantDAO restaurantDAO) {this.restaurantDAO = restaurantDAO;};
+    public RestaurantService(@Qualifier("postgres") RestaurantDAO restaurantDAO) {
+        this.restaurantDAO = restaurantDAO;
+    };
 
     public void addRestaurant(Restaurant restaurant){
+
         restaurantDAO.addRestaurant(restaurant);
+
     };
 
     public void removeRestaurant(long id){
+
+        if (restaurantDAO.getRestaurantName(id).isEmpty())
+        {
+            throw new ResourceNotFound("Restaurant with this id not found");
+        }
+
         restaurantDAO.removeRestaurant(id);
     };
 
     public void updateRestaurant(long id, Restaurant restaurant){
+
+        if (restaurantDAO.getRestaurantName(id).isEmpty())
+        {
+            throw new ResourceNotFound("Restaurant with this id not found");
+        }
+
         restaurantDAO.updateRestaurant(id, restaurant);
     };
 
     public Optional<Restaurant> getRestaurantName(long id){
+
+        if (restaurantDAO.getRestaurantName(id).isEmpty())
+        {
+            throw new ResourceNotFound("Restaurant with this id not found");
+        }
+
         return restaurantDAO.getRestaurantName(id);
     };
 
     public Optional<List<Restaurant>> selectRestaurantFromCriteria(RestaurantCriteria restaurantCriteria) {
+
+        if(restaurantDAO.selectRestaurantFromCriteria(restaurantCriteria).isEmpty())
+        {
+            throw new ResourceNotFound("No restaurants found matching these criteria");
+        }
         return restaurantDAO.selectRestaurantFromCriteria(restaurantCriteria);
     }
 
@@ -39,24 +67,7 @@ public class RestaurantService {
         restaurantDAO.updateRestaurantAverageRating(id, newAverage);
     }
 
-//    public List<Restaurant> filterRestaurants(long restaurant_ID, float averageRating) {
-//        List<Restaurant> ids = new ArrayList<>();
-//        List<Restaurant> ratings = new ArrayList<>();
-//
-//        for (int i=0; i < ids.length; i++) {
-//            if (// if restaurant id 1 is greater > equal then takes over)
-//        } ids.add(restaurant_ID);
-//        averageRating.add(restaurant);
-//    }
-//    int listOfRating = ids.size();
-//    List<Restaurant> idx = new ArrayList<>(listOfRating);
-//    for (int i=0; i < listORating; i++) {
-//        idx.add("add rating of restaurant");
-//    }
-//    sort out ratings of code
-
-    public List<Restaurant> sortRestaurantsByRatings(){
-      return restaurantDAO.sortRestaurantsByRatings();
+    public List<Restaurant> getAllRestaurants() {
+        return restaurantDAO.getAllRestaurants();
     }
-
 }
